+++ dex/pair/src/pair_actions/mod.rs
@@ -4,3 +4,4 @@ pub mod common_result_types;
 pub mod initial_liq;
 pub mod remove_liq;
 pub mod swap;
+pub mod views;
