+++ dex/pair/src/pair_actions/views.rs
@@ -1,0 +1,113 @@
+use crate::{ERROR_NOT_ENOUGH_RESERVE, ERROR_UNKNOWN_TOKEN, ERROR_ZERO_AMOUNT};
+
+multiversx_sc::imports!();
+
+#[multiversx_sc::module]
+pub trait ViewsModule:
+    crate::liquidity_pool::LiquidityPoolModule
+    + crate::amm::AmmModule
+    + crate::contexts::output_builder::OutputBuilderModule
+    + crate::locking_wrapper::LockingWrapperModule
+    + crate::events::EventsModule
+    + crate::safe_price::SafePriceModule
+    + crate::fee::FeeModule
+    + crate::config::ConfigModule
+    + token_send::TokenSendModule
+    + permissions_module::PermissionsModule
+    + pausable::PausableModule
+    + super::common_methods::CommonMethodsModule
+{
+    #[view(getTokensForGivenPosition)]
+    fn get_tokens_for_given_position(
+        &self,
+        liquidity: BigUint,
+    ) -> MultiValue2<EsdtTokenPayment<Self::Api>, EsdtTokenPayment<Self::Api>> {
+        self.get_both_tokens_for_given_position(liquidity)
+    }
+
+    #[view(getReservesAndTotalSupply)]
+    fn get_reserves_and_total_supply(&self) -> MultiValue3<BigUint, BigUint, BigUint> {
+        let first_token_id = self.first_token_id().get();
+        let second_token_id = self.second_token_id().get();
+        let first_token_reserve = self.pair_reserve(&first_token_id).get();
+        let second_token_reserve = self.pair_reserve(&second_token_id).get();
+        let total_supply = self.lp_token_supply().get();
+        (first_token_reserve, second_token_reserve, total_supply).into()
+    }
+
+    #[view(getAmountOut)]
+    fn get_amount_out_view(&self, token_in: TokenIdentifier, amount_in: BigUint) -> BigUint {
+        require!(amount_in > 0u64, ERROR_ZERO_AMOUNT);
+
+        let first_token_id = self.first_token_id().get();
+        let second_token_id = self.second_token_id().get();
+        let first_token_reserve = self.pair_reserve(&first_token_id).get();
+        let second_token_reserve = self.pair_reserve(&second_token_id).get();
+
+        if token_in == first_token_id {
+            require!(second_token_reserve > 0u64, ERROR_NOT_ENOUGH_RESERVE);
+            let amount_out =
+                self.get_amount_out(&amount_in, &first_token_reserve, &second_token_reserve);
+            require!(second_token_reserve > amount_out, ERROR_NOT_ENOUGH_RESERVE);
+            amount_out
+        } else if token_in == second_token_id {
+            require!(first_token_reserve > 0u64, ERROR_NOT_ENOUGH_RESERVE);
+            let amount_out =
+                self.get_amount_out(&amount_in, &second_token_reserve, &first_token_reserve);
+            require!(first_token_reserve > amount_out, ERROR_NOT_ENOUGH_RESERVE);
+            amount_out
+        } else {
+            sc_panic!(ERROR_UNKNOWN_TOKEN);
+        }
+    }
+
+    #[view(getAmountIn)]
+    fn get_amount_in_view(&self, token_wanted: TokenIdentifier, amount_wanted: BigUint) -> BigUint {
+        require!(amount_wanted > 0u64, ERROR_ZERO_AMOUNT);
+
+        let first_token_id = self.first_token_id().get();
+        let second_token_id = self.second_token_id().get();
+        let first_token_reserve = self.pair_reserve(&first_token_id).get();
+        let second_token_reserve = self.pair_reserve(&second_token_id).get();
+
+        if token_wanted == first_token_id {
+            require!(
+                first_token_reserve > amount_wanted,
+                ERROR_NOT_ENOUGH_RESERVE
+            );
+
+            self.get_amount_in(&amount_wanted, &second_token_reserve, &first_token_reserve)
+        } else if token_wanted == second_token_id {
+            require!(
+                second_token_reserve > amount_wanted,
+                ERROR_NOT_ENOUGH_RESERVE
+            );
+
+            self.get_amount_in(&amount_wanted, &first_token_reserve, &second_token_reserve)
+        } else {
+            sc_panic!(ERROR_UNKNOWN_TOKEN);
+        }
+    }
+
+    #[view(getEquivalent)]
+    fn get_equivalent(&self, token_in: TokenIdentifier, amount_in: BigUint) -> BigUint {
+        require!(amount_in > 0u64, ERROR_ZERO_AMOUNT);
+        let zero = BigUint::zero();
+
+        let first_token_id = self.first_token_id().get();
+        let second_token_id = self.second_token_id().get();
+        let first_token_reserve = self.pair_reserve(&first_token_id).get();
+        let second_token_reserve = self.pair_reserve(&second_token_id).get();
+        if first_token_reserve == 0u64 || second_token_reserve == 0u64 {
+            return zero;
+        }
+
+        if token_in == first_token_id {
+            self.quote(&amount_in, &first_token_reserve, &second_token_reserve)
+        } else if token_in == second_token_id {
+            self.quote(&amount_in, &second_token_reserve, &first_token_reserve)
+        } else {
+            sc_panic!(ERROR_UNKNOWN_TOKEN);
+        }
+    }
+}
